ml-training/SmartSynch/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
sentence-transformers==2.2.2
torch==1.9.0
python-dotenv==0.19.0
redis==3.5.3
requests==2.26.0
numpy==1.21.2
scikit-learn==0.24.2
huggingface-hub==0.4.0
transformers==4.18.0
